Rill Erodibility changer
Python Script that changes rill erodibility in WEPP soil files. It can also be tailored to change any other parameters with a few tweaks. 
Mariana did most of it and then I changed a few things to fix some of the bugs.
( Let me know if there are things in it that should be changed to fit coding conventions/ make it better)